refactor(hooks): reorganize change tracking and agent enforcer logic
- Removed deprecated `trackChange` export and streamlined hook exports for clarity.
- Introduced utility functions for change logging and path classification to enhance modularity.
- Moved `buildLangTag` function to a dedicated module for better organization.
- Cleaned up the `setup` module by extracting the `hasIntentMdInTree` function into a separate utility file.
- Improved code maintainability and readability across various hook modules.
