+++ packages/filid/src/hooks/index.ts
@@ -1,13 +1,9 @@
 // barrel -- re-exports all public APIs
 
 export * from './pre-tool-use/pre-tool-use.js';
-export * from './intent-injector/intent-injector.js';
-export * from './pre-tool-validator/pre-tool-validator.js';
-export * from './structure-guard/structure-guard.js';
 export * from './agent-enforcer/agent-enforcer.js';
 export * from './context-injector/context-injector.js';
 export * from './change-tracker/change-tracker.js';
 export * from './session-cleanup/session-cleanup.js';
 export * from './setup/setup.js';
 export * from './shared/shared.js';
-export * from './user-prompt-submit/user-prompt-submit.js';
